+++ CHANGELOG.md
@@ -1,12 +1,39 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r233-release...9996f97ec00718c7f4a643a6fb4702dcfd806b96)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r234-release...de8123b2f3aa79a5b61077af1e19c5d2130ce54a)
 
-- Fix GetSpellCharges
-- Fix wrath buff tracker
-- Fix empty expansion
-- Prevent nil indicator error
-- Increase frame level of AoE Healing
-- Add position for ReadyCheckIcon
-- OverrideLGF by default
-- Update range check (not fix)
-- Update README_zhCN.md
-- Update locales
+Vollmer:
+
+- Add Show Raid toggle (#176)
+
+honeyhoney:
+
+- Fix Full Health Color (#175)
+
+enderneko:
+
+- Fix "Flash" animation
+- Fix appearance preview
+- Fix indicator group size
+- Fix QuickAssist
+- Fix raid debuff creation on classic
+- Fix CheckVehicleRoot
+- Fix stack visibility
+- Fix icon aspect ratio
+- Fix dispellableByMe option
+- Fix BR Timer
+- Fix SR/DR
+- Update gradient health color options
+- Update CUSTOM_CLASS_COLORS
+- Update frame layers of QuickAssist
+- Update header alignment
+- Update CURRENT_SEASON_INDEX
+- Update indicator default frame levels
+- Make custom externals/defensives track by id
+- Ensure custom indicators are loaded correctly
+- AuraTooltips will not block buttons in some cases
+- Add Cell.GetUnitFramesForLGF
+- Add function to adjust StatusText alignment
+- Add expansion data for ruRU
+- Implement "Border" indicator type
+- Init raid debuffs for tww
+- Remove "OnHide" from Glow
+- Remove OverrideLGF